feat: send templated email when COJ received
When a COJ is received an event will be emitted containing the details
of the programme membership and signed COJ.
Listen for COJ received events and send a personalised email to
notification the user that their COJ was received.

The email should address `Dr <Surname>` when possible, otherwise `Dear
Doctor` should be used.
If available, the LO and received date should be included in the email.
A link should be provided to give easy access to TSS for download of PDF
copies - it is not possible to route them directly to the COJ view due
to application state, so they will be sent to the programmes tab
instead.

TIS21-5058
TIS21-5055