feat: get user details from Cognito account
Create a UserAccountService which will provide methods for retrieving
the user account ID and details for a given TIS person/trainee ID.
As the person ID is stored in a custom attribute, the results will be
cached for 24 hours due to the inability to filter by custom attributes
when querying Cognito.

Update the COJ event handling to retrive the trainee's email and name to
use for the email messaging.

TIS21-5058
TIS21-5055