test: mock region retrieval
The CognitoConfiguration tests are failing because, while the region
provider is mocked there are no expectations set. As a result, `null` is
returned for `AwsRegionProvider.getRegion()` and the default builder
behaviour is being used - i.e. checking ENVs, properties etc.
Return a static region when the mocked region provider is called.

TIS21-5058
TIS21-5055