fix: COJ notification formatting
The Conditions of Joining notification is not correctly formatted in all
cases. If the `syncedAt` timestamp is not available there is an
additional space before the period of the event detail.

> We want to inform you that your local deanery office has received you
> signed conditions of joining .

Another issue was the managing deanery and name appearing incorrectly
when empty strings rather than `null` values.

e.g.

> Dear Dr ,

> We want to inform you that your local deanery office () has received
your signed Conditions of Joining.

Update the COJ notification template to ensure that whitespace is used
correctly and more variables handle empty inputs as well as null.

Refactor ConditionsOfJoiningListener to just provide the family name
to the template instead of pre-transforming it.

Write integration tests to verify the behaviour of the template.
ConditionsOfJoiningListenerIntegrationTest should verify each
combination of available template variables specific to Conditions of
Joining notifications.
EmailServiceIntegrationTest should test that all email templates use a
consistent style e.g. greeting.

TIS21-5115
TIS21-5134