refactor: move listener functionality to service
The ConditionsOfJoiningListener performs user account lookup and
localisation, which should be moved to the EmailService so the behaviour
is shared by other sent emails.
Refactor the implementation and tests to move this behaviour.

Refactor the coj-confirmation.html to extract the greeting in to a new
greeting.html fragment, which can be shared across templates.

TIS21-5115
TIS21-5134