feat: store a notification history after sending
When emails are sent, a history of the notification should be stored in
the database.
The history record should contain everything needed to reconstruct the
notification at a later date.

The notification history is the last piece that cannot be implemented or
recovered at a later date. As such, the version should be bumped to
`1.0.0` as the first full MVP.

TIS21-5265
TIS21-5124