feat: ability to view notification messages
Add a new endpoint at `/api/history/message/{id}` to allow the message to
be retrieved for a given notification ID.
Move the existing `/api/history/{id}` endpoint to
`/api/history/trainee/{id}` to avoid any confusion with the new
endpoint.

Update HistoryService to find the notification history and rebuild the
message from the given template name and variables.

Update the TemplateService to handle localization of `java.util.Date`
objects, which is the data type used when restoring a timestamp from the
database.

Update the email templates to include headers for formatting the

Refactor duplicated parameterized test method sources in to a new test
utility class.

Update Java to version 21.

TIS21-5124
TIS21-5266