build: set workflow Java versions
Java 21 is not available during the workflow as the default is 17, Java
21 needs to be passed to the workflow to allow the build to complete.

Gradle 8.4 supports building FOR Java 21, but not running WITH Java 21.
So we must also make Java 17 available in the workflow.

The last provided version will be used by default, so `21, 17` is the
preferred order to ensure that Gradle gets a compatible version to build
with.

TIS21-5124