feat: add notification ID to email headers
We need to be able to relate failed emails to the captured notification
history. To do so, the notification history ID should be sent in the
email headers so that it can be read later.

Update the EmailService to generate the ObjectId for the history and add
it to the email message's headers.

Fix an issue in the UserId cache config that causes empty results to be
cached, this resulted in IDs not being matched if they have recently
been updated. This should never happen in the real world, but could
easily cause issues in dev/stage environments when testing.

TIS21-5444
TIS21-5426