feat: update history when emails fail
When an email bounces or a complaint if received an event is emitted by
SES which ends up on the email failure queue.
Create an event listener, which consumes messages from this queue and
updates the stored history to reflect the failure status.

Create a migrator to append a status of `SENT` to all previous
notification history documents. This is most likely false, as some of
these emails will have failed. However, without additional data this is
the best assumption we can make.

TIS21-5445
TIS21-5426