feat: allow trainees to update notification status
Trainees should be able to update their notification status to mark
in-app notifications as read, unread or archived.
This should be implemented as individual endpoints so that they are
restricted on what the status can be set to and removing the possibility
of modifying any other fields.

When the notification is marked as read, a new `readAt` field should
record the timestamp. Other status changes will not modify that field
further and will not be separately recorded. Only the initial reading of
the notification matters for data purposes.

TIS21-5682
TIS21-5686