feat: create indemnity insurance notification
The indemnity insurance notification should be sent to all new starters
taking part in the notification pilot.
There is a conditional section that should be shown dependant on the
specialty being in a block indemnity scheme.

Update the Curriculum to include `curriculumSpecialtyBlockIndemnity`.
When a qualifying PM is detected, check all associated curricula for any
in a block indemnity scheme. When found then show the conditional
notification section.

TIS21-5617
TIS21-5837