fix: repeated caching of missing users
When a trainee ID is not found in the existing cache, a lookup is
performed on the Cognito user pool and all users are re-cached. As there
can be a large number of person records without Cognito accounts it can
lead to attempts to constant re-cache the user pool over and over.

A fifteen minute window should be added after each full cache operation,
during which no re-caching will occur. This will help reduce demand on
the Cognito API when processing large numbers of contact detail changes.
The fifteen minutes is fairly arbitrary and may need tweaking, until a
longer term solution can be put in place.

TIS21-5904