feat: send to profile email address
The trainee profile was not explicitly trusted and is only used when no
Cognito account was found based on the trainee ID. The Cognito looking
is done using an expensive caching operation as there is no direct way
to query based on tranine ID.

Now that the Cognito account and trainee profile emails are kept in sync
it is preferred to lookup the Cognito account based on the email from
the trainee profile.

Update the NotificationService to retrieve the trainee profile first,
then perform a single Cognito query to find an existing account for the
email address.
There is a small edge case where the Cognito account could not be
updated in time either due to either race conditions or issues like
multiple accounts (in which case the UM service errors), in those cases
the email will be sent based on the profile email instead of erroring
out - and may be flagged as "not registered" incorrectly.

TIS21-5967
TIS21-5969