feat: store failure when email missing
Currently when an email is missing the sending causes an error and the
triggering message ends up on the DLQ. This means that Local Offices are
unable to report on these particular types of failure easily.

Update the NotificationService and EmailService to handle `null` or
empty values for the email field on the trainee record.
If the email is `null` or empty then no attempt should be made to find
the associated user account.

TIS21-5984
TIS21-5985