fix: resending failures when previous contact null
When an email address is not available during sending, the failure is
recorded with no email address.
Upon receiving an updating set of contact details the previously failed
message should be resent.

Update the ContactDetailsService to properly handle the `null` value is
when determining whether to resend the message or not.

TIS21-5984