feat: notify trainees of email changes
When a user account email is changed, the trainee should receive an
email to both the new and old email addresses to inform them of the
change to their contact and sign-in details.

Update the UserAccountListener to receive account update events and use
the EmailService to send both emails.

TIS21-6006
TIS21-6009
TIS21-6010