fix: get user details without increasing MAU
Cognito is charged based on monthly active users, this is usually a
fairly low number and within free limits. However, it was found that the
usage of `AdminGetUser` causes the requested account to be marked as
active for that month.
After reloading all placement/PM data to trigger notification
generation, that resulted in 66k active users. Most of which were not
actually active.

Update the UserAccountService to use a filtered `listUsers` request
instead of the `AdminGetUser` request. The only difference between the
results is some minor structure changes and some MFA attributes not
being included in `listUsers`. For our case both of these differences
are acceptable.
The `sub` and `email` are both unique, so there is no need to handle
multiple results.

TIS21-6103