fix: ignore COJ publish if confirmation sent
The COJ listener previous used COJ received event, which was only ever
sent once. Now that it used a COJ published event, there is a
possibility that the event could be sent again if a PDF was regenerated.

Update the listener to ensure that processing is skipped if a COJ
confirmation was already sent.

TIS21-6106
TIS21-6122