fix: LocalOfficeContact class issue
The call to the restTemplate to retrieve the local office contacts was
not being mapped to a Set<LocalOfficeContact> but a Set<HashMap<>>. As I
understand it, Jackson needs to be configured to do this, but nothing I
could arrange in terms of mappers and config seemed very elegant or
functional.

I ended up with something like
public Set<LocalOfficeContact> getTraineeLocalOfficeContacts(String personId,
      LocalOfficeContactType contactType) {
    try {
      ResponseEntity<LocalOfficeContact[]> loContactsResponse
          = restTemplate.getForEntity(serviceUrl + API_TRAINEE_LOCAL_OFFICE_CONTACTS,
          LocalOfficeContact[].class, Map.of(TIS_ID_FIELD, personId, CONTACT_TYPE_FIELD,
              contactType));
      if (loContactsResponse != null && loContactsResponse.getBody() != null) {
        return Arrays.stream(loContactsResponse.getBody()).collect(Collectors.toSet());
      } else {
        log.warn("Empty body from local-office-contacts endpoint for trainee {} type {}.",
            personId, contactType);
      }
    } catch (RestClientException rce) {
      log.warn("Exception requesting local-office-contacts endpoint for trainee {} type {}: {}",
          personId, contactType, rce.toString());
    }
    return Collections.emptySet();
  }
before just deciding to bin the LocalOfficeContact record class
altogether.

The notification service uses 'raw' List<Map<String, String>> for the
owner contacts functionality, so there is a precedent :/.

TIS21-6669