chore: create local profile properties
Create `application-local.yml` with opinionated defaults to allow the
service to be started without any additional configuration and custom
environmental variables.

There is still an expectation that Localstack resources will be created
upfront as referred to in the properties file.

Bump Spring Boot patch version to resolve an issue with generating
docker images on 3.3.0

NO-TICKET