refactor(ltft): LTFT update event DTO
Refactor the update event DTO to flatten the properties to a simpler
structure.
The whole event DTO should be passed in the template variables and the
template updated to pull from the event DTO, rather than manually set
fields with inconsistent naming.

Rewrite LtftListenerIntegrationTest to be trigger by SQS publishing
instead of calling the listener method directly.
This allows proper verification of the incoming event structure and the
conversion to the newly simplified flatter event structure.
This change also accounts for the structural changes to the published
event, which removed the `content` wrapper around several properties.

TIS21-7178
TIS21-7183