test: fix LTFT integration tests for multiple mail
The `@BeforeEach` set-up was mocking `mailSender.createMimeMessage()`
and returning `new MimeMessage(null)`, because `mock.thenReturn()` will
give the same exact response for every invocation each MimeMessage
created by the mail sender was the same instance.
This led to the captured MimeMessage being overwritten/duplicated when
multiple mails are expected.

Change the set-up to use `mock.thenAnswer()`, which is re-evaluated on
each invocation - resulting in a new instance of MimeMessage each time.

TIS21-6596
TIS21-7215