fix: bump notifications window to 24 hours
A notification window exists which spreads notification scheduling out
over 9 hours to try and avoid rate limits accessing the Cognito API.
However, we are now seeing delivery failures to gmail accounts as a
result of Google blocking our sudden surge in emails sent to them.

Increase the notification window to 24 hours in an attempt to smooth the
surges and avoid the automated blocking.
A longer term fix remains to buffer scheduled emails in to a queue,
which can then be dynamically delayed/retried more easily based on these
sorts of bounces.

TIS21-7290