fix: require notifications to be scheduled to send
Sending from the outbox should require a notification to be scheduled.
This will avoid any duplicate notifications if a schedule is added to
the queue multiple times, as the first attempt will change the status.

This could potentially be opened up to also allow failure retries in the
future, but for now the only path through the outbox is the cron job
adding scheduled notifications.

TIS21-7461
TIS21-7465