feat: add xray tracing for SQS listeners
Add the ability to instrument SQS listeners, the incoming trace header
will be used to determine whether tracing should be sampled or not.

Update AwsXrayInterceptor to add a pointcut for `XRayEnabled`
`SqsListeners` and then trace around it. The SQS listener must use
`Message<T>` as the incoming parameters in order to make the message
headers available to the interceptor.
The interceptor should handle cases where a different parameter is used,
or the trace header is missing and should not enable a sampled trace -
but must still start a segment to avoid errors from downstream services
marked with `@XRayEnabled`.

Update OutboxListener to use a `Message<ObjectIdWrapper>` parameter and
annotate the class with `@XRayEnabled`.

TIS21-7461
TIS21-7465