refactor: remove quartz scheduler
The quartz scheduler is currently disabled and has been replaced with a
MongoDB and Cron based technique, which avoids the duplication of
scheduled notifications between two different databases and allows
better control over intelligent throttling and retry mechanisms.

The remaining quartz code should be refactored and the dependencies
removed from the build.

The `InsertScheduledEmailHistory` migrator needs to remain to avoid
start-up errors, but as it solely interacted with quartz scheduling the
migrator code should be removed and replaced with a warning. This
migrator should never run in on the live environment.

Any still existing Quartz tables will be dropped via migrator.

TIS21-7463
TIS21-7688
TIS21-7689